refactor(lib): split `addStyles` into parts
