+++ lib/utils/applySingleton.js
@@ -1,0 +1,30 @@
+var replaceText = (function () {
+  var textStore = [];
+
+  return function (index, replacement) {
+    textStore[index] = replacement;
+
+    return textStore.filter(Boolean).join('\n');
+  };
+})();
+
+function applyToSingleton (style, index, remove, obj) {
+  var css = remove ? "" : obj.css;
+
+  if (style.styleSheet) {
+    style.styleSheet.cssText = replaceText(index, css);
+  } else {
+    var cssNode = document.createTextNode(css);
+    var childNodes = style.childNodes;
+
+    if (childNodes[index]) style.removeChild(childNodes[index]);
+
+    if (childNodes.length) {
+      style.insertBefore(cssNode, childNodes[index]);
+    } else {
+      style.appendChild(cssNode);
+    }
+  }
+}
+
+module.exports = applyToSingleton;
