+++ options.json
@@ -1,4 +1,3 @@
-
 {
   "type": "object",
   "properties": {
