+++ url.js
@@ -1,7 +1,3 @@
-/*
-	MIT License http://www.opensource.org/licenses/mit-license.php
-	Author Tobias Koppers @sokra
-*/
 var path = require('path');
 
 var loaderUtils = require('loader-utils');
@@ -10,29 +6,32 @@ var validateOptions = require('schema-utils');
 module.exports = function () {};
 
 module.exports.pitch = function (request) {
-	if (this.cacheable) this.cacheable();
+  if (this.cacheable) this.cacheable();
 
-	var options = loaderUtils.getOptions(this) || {};
+  var options = loaderUtils.getOptions(this) || {};
 
-	validateOptions(require('./options.json'), options, 'Style Loader (URL)');
+  validateOptions(require('./options.json'), options, 'Style Loader (URL)');
 
-	options.hmr = typeof options.hmr === 'undefined' ? true : options.hmr;
+  options.hmr = typeof options.hmr === 'undefined' ? true : options.hmr;
 
-	var hmrCode = [
-		"// Hot Module Replacement",
-		"if(module.hot) {",
-		"\tmodule.hot.accept(" + loaderUtils.stringifyRequest(this, "!!" + request) + ", function() {",
-		"\t\tupdate(require(" + loaderUtils.stringifyRequest(this, "!!" + request) + "));",
-		"\t});",
-		"\tmodule.hot.dispose(function() { update(); });",
-		"}"
-	].join("\n");
+  var hmr = [
+    '// Hot Module Replacement',
+    'if(module.hot) {',
+    '  module.hot.accept(' + loaderUtils.stringifyRequest(this, '!!' + request) + ', function() {',
+    '  update(require(' + loaderUtils.stringifyRequest(this, '!!' + request) + '));',
+    '  });',
+    '  module.hot.dispose(function() { update(); });',
+    '}'
+  ].join('\n');
 
-	return [
-		"// style-loader: Adds some reference to a css file to the DOM by adding a <link> tag",
-		"var update = require(" + loaderUtils.stringifyRequest(this, "!" + path.join(__dirname, "lib", "addStyleUrl.js")) + ")(",
-		"\trequire(" + loaderUtils.stringifyRequest(this, "!!" + request) + ")",
-		", " + JSON.stringify(options) + ");",
-		options.hmr ? hmrCode : ""
-	].join("\n");
+  return [
+    '// Style Loader',
+    '// Adds some reference to a css file to the DOM',
+    '// by adding a <link> tag',
+    'var update = require(' + loaderUtils.stringifyRequest(this, '!' + path.join(__dirname, 'lib', 'link/index.js')) + ')(',
+    '\trequire(' + loaderUtils.stringifyRequest(this, '!!' + request) + ')',
+    ', ' + JSON.stringify(options) + ');',
+    '',
+    options.hmr ? hmr : ''
+  ].join('\n');
 };
